Appends an AssemblyVersion attribute to AssemblyInfo files
It doesn't check if the AssemblyInfo file already contains an
AssemblyVersion attribute, and thus cannot ignore that file, or remove
the existing attribute. This means that the build would fail.
Ignoring AssemblyInfo files with an AssemblyVersion would probably cause
a lot of issues with forgotten attributes, therefore we should
remove/replace existing AssemblyVersion attributes in a future version.